requirements.txt
pywebview==3.5
pyinstaller